fix(iOS, Tabs): Updated logic for breaking retain cycles (#3099)
## Description

Closes:
https://github.com/software-mansion/react-native-screens-labs/issues/340

In this PR, I'm fixing the issue with invalidating the controller on the
view component when the view was detached from the window. It wasn't the
correct approach, as the view might be reattached with the associated RN
component.

In the current approach, we're observing incoming mutations on mounting
transactions. If the Host component is about to be removed, it asks all
its children to invalidate (remove the reference to the controller), and
then it invalidates itself. In this approach, I assume that all Host
components will be listening to the mounting transactions and will be
responsible for invalidating their children.

Note: There will still be a lot of work to be done. We need to revisit
the assumed conditions here, for other components (e.g. for stack) we'd
need to detect not only the Host component's `Delete` mutation, but also
any `Screen` component's mutation and detach the single screen. What I'd
consider is adding `shouldInvalidateOnMutation` to the protocol and
treating it as a determinant of what and when should be detached; the
logic for detaching might slightly differ between components. I think
that it'd be better to prepare a skeleton with a POC for a single
component here and then extend other components with this protocol in
separate PRs.

## Changes

- Added a protocol and registry responsible for view invalidation
- Updated logic for detecting which components should be invalidated -
now we're relying on `RCTMountingTransactionObserving`

## Test code and steps to reproduce

Tested with the example from:
https://github.com/software-mansion/react-native-screens-labs/issues/340#issuecomment-3158525815

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes