fix(android, Tabs): Fix loading external resources in release mode (#3148)
## Description

In release, assets are stored under
`android/app/src/main/res/[drawable|raw|...]/` and what
`Image.resolveAssetSource` returns is the name of the Image, which is
stored in the path above. In that case, loading icons is failing, as our
`ImageLoader` doesn't recognize this string. For fixing it, I'm adding a
support for finding resources using `context.resources.getIdentifier`
API.

## Changes

- Added some methods which are repsonsible to differentiate between
debug and release resources.

## Test code and steps to reproduce

Test with `TestBottomTabs` in release.

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes