feat!(iOS, Tabs): support for appearance types, layouts and states (#3078)
## Description

Adds support for appearance types, layouts and states in iOS BottomTabs.

While working on this PR, I noticed a bug on Paper
https://github.com/software-mansion/react-native-screens-labs/issues/326.
It is related to this change (now we allow to set scrollEdgeAppearance
to `nil`) but I think it should be handled separately.

## Changes

- full support for `standard, scrollEdge` appearance, `stacked, inline,
compactInline` layouts, `normal, selected, focused, disabled` states
- send apperance props via `UnsafeMixed` to avoid codegen issues (nested
objects are set to null, no optionals) or having 192 props inside
component (if we wanted to flatten the hierarchy)
- make all appearance configurations independent of each other (map 1:1
to native UIKit behavior, without inheritance) - this was done mostly
due to some bugs in UIKit/SwiftUI with handling appearance from many
appearance configs


https://github.com/user-attachments/assets/26d53af0-341f-4c96-b2cd-1cf622404b02

- remove appearance config from host component

## Test code and steps to reproduce

Run `TestBottomTabs`.

## Checklist

- [x] Included code example that can be used to test this change
- [ ] Ensured that CI passes

---------

Co-authored-by: Tomasz Boroń <tomasz.boron@swmansion.com>