fix(JS): Keep relative imports for backward compatibility (#3153)
## Description

In this commit:
https://github.com/software-mansion/react-native-screens/commit/3a3a6c5449796aeef229d40b5805721b808bf3ab
`eslint` was complaining about the import. Moreover, running `lint-js`
is autofixing this line itself after upgrading to 0.81. As for now, we
need to keep relative import for a backward compatibility with versions
that we support and for which `codegenNativeComponent` wasn't exported.

Fixes
https://github.com/software-mansion/react-native-screens/issues/3152 .

## Changes

- Skipping linter check for the line with the relative import.

## Test code and steps to reproduce

- Create a project from template, using for example 0.79.6
- Add `react-native-screens@^4.15.0` and `@react-navigation` 
- Add native stack to `App.tsx`
- Run `yarn test`

For testing my changes I packed RNScreens from this PR to tarball and
sideloaded it to the application with `file://`

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes