fix(Android): add custom fragment factory & modify library installation steps to prevent on-restoration crashes (#3089)
## Description

<!--
Description and motivation for this PR.
Include Fixes #<number> if this is fixing some issue.

Fixes # .
-->

This PR replaces the current workaround for crashes connected to the
configuration changes on Android. The previous fix required disabling
passing saved instance state in Activity onCreate - `onCreate(null)`.
Such a solution generates a problem with state persistence on the native
side of brownfield applications. Because of that, we decided to replace
it with a custom fragment factory that will limit the scope of the
solution and fix the problem directly only on the React Native side. All
fragments connected to the RN Screens package will be replaced with
self-destructing ones, which will prevent Android from recreating the
Fragment after the configuration change. We will leave this
responsibility for React Native. Thanks to that, there won't be any
conflict between the Android system and React Native, and we will
prevent the app from crashing or any other app appearance-related
issues.

Credits to @bojanlozanovski77 for bringing up the idea in
https://github.com/software-mansion/react-native-screens/discussions/2917

Fixes:
https://github.com/software-mansion/react-native-screens/issues/17

## Changes

<!--
Please describe things you've changed here, make a **high level**
overview, if change is simple you can omit this section.

For example:

- Updated `about.md` docs

-->

- Created a custom fragment factory (RNScreensFragmentFactory.kt) that
won't instantiate RN Screens fragments
- Created self-destructing fragments (NoOpFragment.kt) that will be
instanciated by the RNScreensFragmentFactory, to prevent the conflict
that crashes the app
- Updated installation steps in library README
- Applied the fix in `MainActivity` files of `FabricExample` and
`Example` apps

<!--

## Screenshots / GIFs

Here you can add screenshots / GIFs documenting your change.

You can add before / after section if you're changing some behavior.

### Before

### After

-->

## Test code and steps to reproduce

<!--
Please include code that can be used to test this change and short
description how this example should work.
This snippet should be as minimal as possible and ready to be pasted
into editor (don't exclude exports or remove "not important" parts of
reproduction example)
-->
1. Remove ignoring config changes in `AndroidManifest.xml` file 
```diff
<activity
    android:name=".MainActivity"
    android:label="@string/app_name"
-   android:configChanges="keyboard|keyboardHidden|orientation|screenLayout|screenSize|smallestScreenSize|uiMode"
    android:launchMode="singleTask"
    ...
```
2. Launch the example app
3. Verify if the app isn't crashing when causing different types of
configuration changes on Android. For example:
	- Orientation change on devices ≥600dp with portrait lock
	- Opening/closing foldables or flip phones  
	- Split/Multi-Window mode  
	- Desktop windowing on tablets  
	- Launching external activities (e.g. system Photo Picker)  
- Changing accessibility settings, permissions, system fonts,
wallpapers, or language

## Checklist

- [X] Included code example that can be used to test this change
- [X] Updated TS types
- [X] Updated documentation: <!-- For adding new props to native-stack
-->
- [X]
https://github.com/software-mansion/react-native-screens/blob/main/guides/GUIDE_FOR_LIBRARY_AUTHORS.md
- [X]
https://github.com/software-mansion/react-native-screens/blob/main/native-stack/README.md
- [X]
https://github.com/software-mansion/react-native-screens/blob/main/src/types.tsx
- [X]
https://github.com/software-mansion/react-native-screens/blob/main/src/native-stack/types.tsx
- [X] Ensured that CI passes

---------

Co-authored-by: bojanlozanovski77 <bojanlozanovski77@users.noreply.github.com>
Co-authored-by: Kacper Kafara <kacperkafara@gmail.com>