Add support for reacting to Appearance.setColorScheme calls
This was a bit tricky. Our views are not notified in any way that the
configuration got changed, but fortunately our fragments do. Therefore,
I've forwarded configuration changes from `TabScreenFragment` to
`TabsHost`. This was done indirectly, through `TabScreen` &
`TabScreenDelegate` to avoid coupling the components.

When doing so I encountered another problem: the resources of our theme
applied to `ContextThemeWrapper` were not updated. This was done
because, the API that RN uses - `AppCompatDelegate.setDefaultNightMode`
applies theme changes only to top level application context (and
configuration changes are then dispatched). The solution would be to
recreate the `ContextThemeWrapper`, however lifecycle of our fragments &
views is bound to the `ContextThemeWrapper`, therefore we would prefer
to avoid recreation of the whole managed hierarchy.

I've settled with manually applying appropriate theme depending on
uimode provided by the configuration. This solves the problem.