feat!(iOS, Stack v4): change search bar related prop defaults (#3186)
## Description

Changes default values for search bar related props.

### `hideWhenScrolling`
On Fabric, for some reason, default value for `hideWhenScrolling` is
`false`, even though:
- on Paper, it is `true`, because we initialize `_hideWhenScrolling` to
`YES`,
- in `GUIDE_FOR_LIBRARY_AUTHORS`, react-navigation's docs, it is stated
that the default value for this prop is `true`,
- UIKit default is `true`.

This PR fixes this.

### `placement`
To preserve backward compatibility when adding `placement` as a prop,
the default value for `placement` became `stacked`. Starting from iOS
26, new search bar placements has been added (`integrated`, ...). In
order to make it even easier for users to get *native feel* in their
apps, we decided to change the default value of this prop to
`automatic`.

> [!WARNING]
>
> This seems like a **BREAKING** change from users perspective, but we
consider it just as a alignment with native behaviour - iOS 26 just
changes things.

## Changes

- change default value of `hideWhenScrolling` to `true` on Fabric
- change default value of `placement` to `'automatic'`
- adjust test screens to use previous default values (I tried to use
`hideWhenScrolling` where necessary)

## Test code and steps to reproduce

Run tests related to search bar.

## Checklist

- [x] Included code example that can be used to test this change
- [ ] Updated TS types
- [ ] Updated documentation: <!-- For adding new props to native-stack
-->
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/guides/GUIDE_FOR_LIBRARY_AUTHORS.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/native-stack/README.md
- [x]
https://github.com/software-mansion/react-native-screens/blob/main/src/types.tsx
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/native-stack/types.tsx
- [ ] Ensured that CI passes

---------

Co-authored-by: Kacper Kafara <kacper.kafara@swmansion.com>