fix(iOS): Prevent delay when dismissing transparentModals (#3189)
## Description

Closes #3074.

This PR removes the delay when dismissing fullscreen modal-like screens.
This behavior can be observed on native apps, but makes things look as
if they are broken. We also observe that modal screens don't take
transitionDuration into account. We don't plan on supporting it now, but
might look into it in the future. Docs are updated to reflect the
current state.

## Testing

Use Test3074. It has 4 screens in stack - 1. and 2. are a regular
screen, 3. & 4. are transparentModals. The last one should dismiss
immediately without animation. Revert changes in RNSScreen to see the
bug. Checking Debug/Slow animations in simulator makes it even more
apparent.

---------

Co-authored-by: Kacper Kafara <kacperkafara@gmail.com>