Move types related to StackScreen to separate file
Conforming to https://github.com/software-mansion/react-native-screens-labs/discussions/313

`maxLifecycleState` is typed so ugly, because before it's type was
dependant on `StackScreenLifecycleState` object (!) which is defined
inside implemetation file & I don't want to place it in types file.
Therefore, to avoid circular import I've typed it manually. I do
acknowledge that its ugly as hell but not aware of better solution atm.