fix(Android): header config shadow node has wrong origin when `translucent == true` (#3239)
## Description

https://github.com/software-mansion/react-native-screens/pull/2466 fixes
Pressable in screen header by shifting the HeaderConfigShadowNode
origin.y by its frame height. (more info in the pull request).
However, setting isTranslucent to true on android displays the header
over the content so we must not apply this correction.

This fixes Pressable for translucent headers on Android.

## Changes

Skip the frame corrections on Android when translucent is true. I'm not
sure that's the cleanest way to fix this issue, but i'm not familiar
enough with react-native-screens to find a better solution. Because this
is an issue I have in my production app and this is critical, I'm
pushing this fix for now.

## Test code and steps to reproduce

https://snack.expo.dev/@dernise/header-navigation-broken

(Run this on a real device by scanning the QR code in expo go, do not
use the emulator provided by appetize)

---------

Co-authored-by: Kacper Kafara <kacperkafara@gmail.com>