fix(Android, Stack): Add logic for reattaching mounting override delegate (#3241)
## Description

Currently, when the app is exited, a new `Surface`, `ShadowTree`, and
`MountingCoordinator` are created due to the destruction and recreation
of the associated Activity (triggered with `onDestroy`). As a result,
the code that invokes `setMountingOverrideDelegate` is never executed
again, which prevents `notifyScreenRemoved` from being called. This
leads to an inconsistent state during screen transitions.

In this PR, I added a listener for the `onHostResume` lifecycle event.
However, since the order of lifecycle events is not guaranteed, we
cannot assume that `onHostDestroy` occurs before `onHostResume`.
Therefore, we cannot mark which resume callbacks should trigger
setupFabric.

> [!CAUTION]
> As a workaround, I'm storing a weak_ptr to the `MountingCoordinator`
and, on every `onHostResume`, I check whether the coordinator has
already registered our callback. We initiated an internal discussion
about the long-term solution, which may require additional effort on the
RN side.

Fixes:
https://github.com/software-mansion/react-native-screens/issues/3234

## Changes

- extended ScreensModule with `LifecycleEventListener` to track
`onHostResume` callback
- added set for tracking which coordinators has registered our callback

## Screenshots / GIFs

Here you can add screenshots / GIFs documenting your change.

You can add before / after section if you're changing some behavior.

### Before


https://github.com/user-attachments/assets/4dbc6e52-7c8e-44a4-946e-05b645ad6cea

### After


https://github.com/user-attachments/assets/de132990-fb15-48dc-9bff-cc33f71cfbde

## Test code and steps to reproduce

Tested with the application from
https://github.com/software-mansion/react-native-screens/issues/3234

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes