feat(iOS): Implement ScrollViewProviding protocol (#3238)
Closes
https://github.com/software-mansion/react-native-screens-labs/issues/456

## Description

This PR implements ScrollViewProviding protocol as per internal
discussion. The protocol enables us to be smarter about finding content
ScrollView than simply descending into 0th child. A class that
implements this protocol can decide how to continue the search. To
complement the feature, ScrollViewFinder is updated with new method,
`findContentScrollViewWithDelegatingToProvider` that takes advantage of
the protocol.

## Changes

- Added `ScrollViewProviding` protocol, with implementations for
`RNSScreenStack`, `BottomTabsHostComponentView`. The implementations
continue search directly from their currently selected screens. This has
the desired side-effect of completing the search even if the screen is
not yet mounted in host hierarchy.
- Modified `ScrollViewFinder` to include a method that takes advantage
of the new protocol, while leaving the simple version
- Concrete implementations of the new protocol will be added in future
PRs

## Test code and steps to reproduce

Use Test3212; For now, nested containers may not behave as expected,
because the new "smart" method is not called yet; things will change in
the following PRs