fix: fix custom screen swipe transitions (#3254)
## Description

This PR fixes custom swipe-back transitions. The issue arose because, in
the new version of React Native, they changed the way to obtain the
shadow node wrapper. Since the ref to the screen is a native ref and we
only need this on Fabric, it was possible to simplify the way we obtain
that reference. Additionally, I found a small issue with the propagation
of the ref and fixed it in the `Screen` implementation.

| before | after |
| --- | --- |
| <img width="443" height="939" alt="image"
src="https://github.com/user-attachments/assets/c2c36f3a-6f46-4fdd-b107-a1e536eb8a9f"
/> | <video
src="https://github.com/user-attachments/assets/1352f57e-c789-4735-9e1d-0905ad396a1d"
/> |

Example:
https://github.com/software-mansion/react-native-reanimated/blob/main/apps/common-app/src/apps/reanimated/examples/ScreenTransitionExample.tsx

## Checklist

- [ ] Included code example that can be used to test this change
- [x] Updated TS types
- [ ] Updated documentation: <!-- For adding new props to native-stack
-->
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/guides/GUIDE_FOR_LIBRARY_AUTHORS.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/native-stack/README.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/types.tsx
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/native-stack/types.tsx
- [ ] Ensured that CI passes