Revert changes to `react-navigation`
This is unfortunately necessary & will break examples for little while,
before `react-navigation` releases new version. It is necessary, because
otherwise people might not be able to check out screens main, as before
this change the submodule commit points to a fork.

To keep our example apps working, I'm commenting the test examples for
this PR temporarily.