feat(iOS): support UIBarButtonItem in header (#2987)
## Description

The current implementation of `headerLeft` and `headerRight` adds a
react view as a custom view in a UIBarButtonItem. This implementation is
sufficient at most times but I believe we can achieve greater "native
feel" if the native stack has an protocol for adding actual
UIBarButtonItems in the header. As the UIBarButtonItems has properties
and features that can be difficult to mimic with a react view.

Also with the introduction of iOS 26, using only custom views in
UIBarButtonItem presents some limitations. Mainly that the adaptive tint
color (based on the underlying view) is not working on a UIBarButtonItem
with a custom view as demonstrated below under "Screenshots".

## Changes

This PR adds the properties `headerRightItems` and `headerLeftItems` on
the native stack Screen that makes it possible to add one or several
UIBarButtonItem to the right/left of the header and/or functions
returning a React Node. Most of the features of the UIBarButtonItem is
supported (see either "Bar Button Items" in the example apps or the
[type
definition](https://github.com/johankasperi/react-navigation/blob/5e120f1aee81ac23375700d43ca8dc2da827c3a5/packages/native-stack/src/types.tsx#L694)).

## Screenshots / GIFs

#### Non adaptive tint color when using old property `headerRight` on
iOS 26

https://github.com/user-attachments/assets/194c70b5-7492-48df-aa2c-8fb889ece461

#### Adaptive tint color when using new property
`headerRightBarButtonItems` on iOS 26

https://github.com/user-attachments/assets/9acb1981-3461-4f28-8a0d-54ea9956d3c0

#### UIBarButtonItem with style "prominent" on iOS 26
![Simulator Screenshot - iPhone 16 Pro - 2025-06-27 at 16 28
39](https://github.com/user-attachments/assets/e674d295-928b-4a97-bed3-17c86a1f541d)

#### UIBarButtonItem with UIMenu on iOS 26

https://github.com/user-attachments/assets/db2f7bab-3ade-423f-b80b-9c35b6cee578

## Test code and steps to reproduce

I've created a screen named "Bar Button Items" in the example app that
showcases all of the proposed features.

## Checklist

- [x] Included code example that can be used to test this change
- [x] Updated TS types
- [x] Updated documentation: <!-- For adding new props to native-stack
-->
- [x]
https://github.com/software-mansion/react-native-screens/blob/main/guides/GUIDE_FOR_LIBRARY_AUTHORS.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/native-stack/README.md
- [x]
https://github.com/software-mansion/react-native-screens/blob/main/src/types.tsx
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/native-stack/types.tsx
_Unsure if I need to do this. I'm targeting react-navigation v6 and
above so no JS changes has been made to react-native-screens in this PR.
Would appreciate some guidance to if and where I should make
documentation changes if needed. Thank you!_
- [ ] Ensured that CI passes

---------

Co-authored-by: Kacper Kafara <kacperkafara@gmail.com>