refactor(iOS): rename `label` -> `title` in context of header items API (#3318)
## Description

This PR intends to align back prop naming with the
[platform](https://developer.apple.com/documentation/uikit/uibaritem/title?language=objc)
nomenclature.

The header item API is iOS specific and there is really no reason to
obfuscate in any way
what we expose. 

## Changes

Renamed `label` -> `title` & `labelStyle` -> `titleStyle`.

I recognize that technically this is a breaking change, however since
react-navigation
has not published the API yet, and we're in coordination, API end-users
won't be affected.

## Test code and steps to reproduce

`BarButtonItems` example should work the same.

## Checklist

- [ ] Ensured that CI passes