refactor!(iOS): move `changesSelectionAsPrimaryAction` to menu button item (#3331)
## Description

This prop has no effect when used with `HeaderBarButtonItemWithAction`,
as it is supposed to change selection value in context of menu button.

It is possible that there are some more behaviours introduced by this
prop, but not spotted yet.



https://github.com/user-attachments/assets/4b4e52bd-80f6-4613-943b-4479a849e732



## Test code and steps to reproduce

`BarButtonItems` example in our app. Add this prop to a menu button
item.

## Checklist

- [ ] Ensured that CI passes