feat(iOS, SplitView): Add support for primaryBackgroundStyle prop (#3320)
## Description

Added support for the primaryBackgroundStyle prop:
https://developer.apple.com/documentation/uikit/uisplitviewcontroller/primarybackgroundstyle?language=objc

This property was previously omitted due to a documentation note stating
it was not supported on iOS. However, while working on a separate task,
I noticed that different values for this prop do produce varying effects
on iOS 26. Therefore, I’ve added support for it accordingly.

## Changes

- Added TS definition
- Added conversions
- Added handling in the appearance coordinator

## Screenshots / GIFs


https://github.com/user-attachments/assets/3b9c81ab-ea04-4306-89cd-9006e054b6f2

## Test code and steps to reproduce

Added new example for testing

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes

---------

Co-authored-by: Krzysztof Ligarski <63918941+kligarski@users.noreply.github.com>