fix(iOS, Tabs): fix problems with dynamically changing title, icon, systemItem (#3333)
## Description

Fixes problems related to updating title and icon dynamically, with or
without systemItem.

If we use `systemItem`, we have 3 possible cases for `title`:
- we want to use title from `systemItem`, (e.g. "Contacts"),
- we want to use custom title,
- we want to use no title (hiding the label).

On Fabric, there is no difference between `undefined` and `''` title.
That's why I added `isTitleUndefined` native prop that will perform this
check for us in JS. On Paper, this problem doesn't happen, which
simplifies the logic a little bit.

Also, I noticed that icons are not reset if we remove them completely -
fixed that as well by setting image to nil if there is no icon name and
`systemItem` is `none`.

## Videos
### Before


https://github.com/user-attachments/assets/78b6a35a-a363-41f1-9e42-f51aa2fe34c0

### After


https://github.com/user-attachments/assets/76ec7f2b-f387-4ea0-ad8b-21d000dfe452

## Changes

- add `isTitleUndefined` prop
- use `systemItem` title if `title == nil`
- match logic on Paper

## Test code and steps to reproduce

Run `TestBottomTabs`. Change icon, title (e.g. `undefined`, `''`,
`'Title'`) and systemItem (e.g. `'contacts'`, `'search'`) and observe
the changes.

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes