test: Refactor Test3212, add option to set config from scroll view screen (#3284)
## Description

This PR includes a small refactor for Test3212 that removes useCallback
in favor of regular functions, and adds a button to change
scrollEdgeEffect config right on the screen it is applied to - to have
more paths to test.

For instance, you can set the scrollEdgeEffect config to 'hard' on the
first Stack screen, push the second one with ScrollView, see the
configuration being applied, then click the button and see the effect
changed to 'hidden'.