feat(iOS, SplitView): Migrate to new invalidate method (#3364)
## Description

SplitViewHost initially was responsible for invalidating children when
the whole structure was destroyed. This approach is no longer necessary,
as every screen can invalidate itself directly when it's deleted.

## Changes

- Removed invalidate from SplitViewHost

## Test code and steps to reproduce

The easiest way to test is to change with fast refresh the number of
columns between 2 and 3 and verifying if the breakpoint in
RNSSplitViewScreenComponentView was hit.

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes