feat(iOS, Tabs): add prop to hide the tab bar (#3343)
## Description

Adds `tabBarHidden` prop to `BottomTabsHostComponentView`.


https://github.com/user-attachments/assets/4b17f5d1-1438-4bf7-b8be-030af4589844

> [!IMPORTANT]
>
> 1. If we hide the tab bar from the beginning and then show it later,
appearance configuration doesn't seem to work on selected tab until you
change the tab and go back. This seems to be a native bug as under
certain circumstances it is reproducible in bare UIKit app but we should
investigate this further. I added a ticket on our board:
https://github.com/software-mansion/react-native-screens-labs/issues/516.
>
> 2. Animating tab bar visibility change is also possible starting from
iOS 18. I decided not to implement this now as we need to prevent
initial animation/think of different API for visibility changes. More
details:
https://github.com/software-mansion/react-native-screens-labs/issues/517.

## Changes

- add `tabBarHidden` prop

## Test code and steps to reproduce

Run `TestBottomTabs`, add `tabBarHidden` prop, observe dynamic changes.

## Checklist

- [x] Included code example that can be used to test this change
- [x] Updated TS types
- [x] Updated documentation: <!-- For adding new props to native-stack
-->
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/guides/GUIDE_FOR_LIBRARY_AUTHORS.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/native-stack/README.md
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/types.tsx
- [ ]
https://github.com/software-mansion/react-native-screens/blob/main/src/native-stack/types.tsx
- [ ] Ensured that CI passes