feat: Bring back gestureResponseDistance on iOS 26 (#3353)
## Description

Closes
https://github.com/software-mansion/react-native-screens-labs/issues/521

We previously removed the support for gestureResponseDistance on iOS 26,
because it was seemingly conflicting with the new native behavior - now,
each screen can be dismissed by swiping on the content, in addition to
swiping from the edge. For the one case of animations, the old gesture
recognizer is used, but to match the new one, it also didn't respond the
the responseDistance.

After the changes, the recognizer is brought back to work as in ios 18,
and the new interactiveContentPop has the check added.

## Test code and steps to reproduce

Use Gestures test -- should work the same way as iOS 18. Use Test3093
with modifications in JS to test custom gestures.
```ts
animation: 'slide_from_bottom',
animationMatchesGesture: true,
// gestureResponseDistance: { ... }
```