feat(iOS, Stack): Add controller invalidating to new stack impl (#3219)
## Description

Added `invalidate` callback from `RCTComponentViewProtocol`.
With the current approach, `invalidate` is triggered as late as
possible, since the component being destroyed is already removed from
the registry by RN, before the `mountingTransactionDidMount` callbacks
are invoked - so it won't receive it.
Therefore, introducing an additional delay (to the next tick) provides a
solution that ensures all transitions have a chance to start before
references are broken.

> [!IMPORTANT]  
> Within this PR, I'm assuming that the minimum supported version for
the stack v5 implementation will be 0.82.

Closes:
https://github.com/software-mansion/react-native-screens-labs/issues/127

## Changes

- Added `invalidate` callback for new stack screen.

## Test code and steps to reproduce

Verified with logs that the callback is triggered and the controller's
reference is removed

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes