feat: add `snapshotAfterUpdates` property to `RNSScreenView` (#3296)
## Description

add `snapshotAfterUpdates` property to `RNSScreenView`. This is done for
the sake of integration with `reanimated` (SETs), where `reaniamted`
needs to make sure that some of the updates are applied before we take
the snapshot.

> [!caution]
> This is not a part of public API, rather an internal one. 

## Test code and steps to reproduce

The default behaviour should be as before - if not modified, the
snapshot should be taken before view updates.

## Checklist

- [x] Ensured that CI passes

---------

Co-authored-by: Kacper Kafara <kacperkafara@gmail.com>