refactor(iOS): extract common image loading code to separate class (#3371)
## Description

Extracts image loading code to `RNSImageLoadingHelper` class. It handles
image templates and synchronous loading in debug for `BarButtonItems`.

Part of
https://github.com/software-mansion/react-native-screens-labs/issues/525.

> [!NOTE]
>
> For this PR I decided to leave custom back button image in
`RNSScreenStackHeaderConfig` as is because it will require more
complicated changes (we render the image and extract the source which
needs to be handled differently than when just receiving the source).

Fixes
https://github.com/software-mansion/react-native-screens-labs/issues/534
(loads `ImageURISource` images with `RCTImageLoader` in debug (->
asynchronously); this causes the image to load after a delay but it's
expected).

## Changes

- extract image loading code to `RNSImageLoadingHelper`
- use `RNSImageLoadingHelper` in `RNSBottomTabsAppearanceCoordinator`
and `RNSBarButtonItem` config

## Test code and steps to reproduce

1. Run `TestBottomTabs`. Verify that icons are present and behavior is
correct (icon 2 should be tinted when tab is selected, icon 3 should
not).
2. Run `BarButtonItems` example screen, open `IconButtonDemo`. Icon
should not be tinted. Change `tinted: false` to `tinted: true`. Icon
should become tinted.

Test on Paper and Fabric, for both debug and release.

## Checklist

- [x] Included code example that can be used to test this change
- [x] Ensured that CI passes